+++ config/service.go
@@ -2,6 +2,7 @@ package config
 
 // Service holds string representations of Service attributes.
 type Service struct {
+	AutoStart      bool `yaml:"auto_start"`
 	Proxy          map[string]string
 	Command        []string
 	StopAfter      string `yaml:"stop_after"`